chore: upgrade letta-code to 0.27.30
